ACCUMULO-4708 Limit RFile block size to 2GB
Fixed redudant casting
Used Preconditions.checkArgument()
Put in a test to check DefaultConfiguration's indexBlockSize

+++ core/src/main/java/org/apache/accumulo/core/data/Key.java
@@ -1221,7 +1221,7 @@ public class Key implements WritableComparable<Key>, Cloneable {
    */
   public boolean sanityCheckKey() {
     // If fails the check
-    if ((long) ((long) this.row.length + (long) this.colFamily.length + (long) this.colQualifier.length + (long) this.colVisibility.length) > Integer.MAX_VALUE) {
+    if (((long) this.row.length + (long) this.colFamily.length + (long) this.colQualifier.length + (long) this.colVisibility.length) > Integer.MAX_VALUE) {
       return false;
     }
     return true;
